changed a few widgets to private widgets since they were only used by one class. preloaded all themes to prevent white screen
