+++ lib/components/game_board/game_board_layout/game_board_layout_large.dart
@@ -21,8 +21,6 @@ class GameBoardLayoutLarge extends StatelessWidget {
   final double puzzleHeight;
   final double puzzlePadding;
 
-  void reset() {}
-
   @override
   Widget build(BuildContext context) {
     AnimeThemeList animeThemeList = context.read<AnimeThemeList>();
