--- lib/components/image_selection/image_selection_components/circle_transition_clipper.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CircleTransitionClipper extends CustomClipper<Path> {
-  final Offset center;
-  final double radius;
-
-  CircleTransitionClipper({
-    required this.center,
-    required this.radius,
-  });
-
-  @override
-  Path getClip(Size size) {
-    return Path()
-      ..addOval(
-        Rect.fromCircle(
-          center: center,
-          radius: radius,
-        ),
-      );
-  }
-
-  @override
-  bool shouldReclip(covariant CustomClipper<Path> oldClipper) {
-    return true;
-  }
-}
